+++ modules/core/src/lib/effect-manager.ts
@@ -1,9 +1,14 @@
 import {deepEqual} from '../utils/deep-equal';
 import {default as LightingEffect} from '../effects/lighting/lighting-effect';
+import type Effect from './effect';
 
 const DEFAULT_LIGHTING_EFFECT = new LightingEffect();
 
 export default class EffectManager {
+  effects: Effect[];
+  _internalEffects: Effect[];
+  _needsRedraw: boolean | string;
+
   constructor() {
     this.effects = [];
     this._internalEffects = [];