+++ modules/core/src/lib/effect.ts
@@ -1,5 +1,10 @@
+import type Layer from './layer';
+
 export default class Effect {
-  constructor(props = {}) {
+  id: string;
+  props: any;
+
+  constructor(props: {id?: string} = {}) {
     const {id = 'effect'} = props;
     this.id = id;
     this.props = {...props};
@@ -7,7 +12,7 @@ export default class Effect {
 
   preRender() {}
 
-  getModuleParameters() {}
+  getModuleParameters(layer: Layer): any {}
 
   cleanup() {}
 }