+++ modules/geo-layers/src/index.ts
@@ -28,3 +28,5 @@ export {default as H3HexagonLayer} from './h3-layers/h3-hexagon-layer';
 export {default as Tile3DLayer} from './tile-3d-layer/tile-3d-layer';
 export {default as TerrainLayer} from './terrain-layer/terrain-layer';
 export {default as MVTLayer} from './mvt-layer/mvt-layer';
+
+export {getURLFromTemplate as _getURLFromTemplate} from './tile-layer/utils';