+++ test/modules/geo-layers/tile-layer/utils.spec.js
@@ -502,6 +502,7 @@ test('urlType', t => {
 
 test('getURLFromTemplate', t => {
   const TEST_TEMPLATE = 'https://server.com/{z}/{x}/{y}.png';
+  const TEST_TEMPLATE2 = 'https://server.com/{z}/{x}/{y}/{x}-{y}-{z}.png';
   const TEST_TEMPLATE_ARRAY = [
     'https://server.com/ep1/{x}/{y}.png',
     'https://server.com/ep2/{x}/{y}.png'
@@ -510,6 +511,11 @@ test('getURLFromTemplate', t => {
     getURLFromTemplate(TEST_TEMPLATE, {x: 1, y: 2, z: 0}),
     'https://server.com/0/1/2.png',
     'single string template'
+  );
+  t.is(
+    getURLFromTemplate(TEST_TEMPLATE2, {x: 1, y: 2, z: 0}),
+    'https://server.com/0/1/2/1-2-0.png',
+    'single string template with multiple occurance'
   );
   t.is(
     getURLFromTemplate(TEST_TEMPLATE_ARRAY, {x: 1, y: 2, z: 0}),